requirements.txt
torch==2.1.0
torchvision==0.16.0
transformers==4.36.0
peft==0.7.0
pillow==10.0.0
pyyaml==6.0